refactor: 관련 컴포넌트 features/registerProjectPost로 이동
